Add deprecated folder for decluttering, rename files
