+++ src/main/java/frc/robot/commands/ChangeCameraPipeline.java
@@ -3,7 +3,7 @@ package frc.robot.commands;
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class ChangePipeline extends Command {
+public class ChangeCameraPipeline extends Command {
 
 	int pipeline;
 
@@ -11,7 +11,7 @@ public class ChangePipeline extends Command {
 	 * Changes the pipeline ID of the limelight, which switches the vision tracking settings.
 	 * Use the static variables in CameraSubsystem.
 	 */
-	public ChangePipeline(int pipeline) {
+	public ChangeCameraPipeline(int pipeline) {
 		requires(Robot.camera);
 		this.pipeline = pipeline;
 	}
