+++ src/main/java/frc/robot/commands/autonomous/deprecated/ApproachCurveVerbose.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.autonomous;
+package frc.robot.commands.autonomous.deprecated;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.util.SynchronousPIDF;
@@ -8,13 +8,12 @@ import frc.robot.QuickAccessVars;
 import frc.robot.Robot;
 import frc.robot.subsystems.CameraSubsystem;
 
-/**
- * ApproachCurve approachs the target keeping it to the left or right (depending
- * on the direction of approach) and when it is aligned up (by the aspect ratio
- * being 2.3 or above), it approach the rest of the distance with it in front.
+/** ApproachCurve approachs the target keeping it to the left or right (depending on the direction
+ * of approach) and when it is aligned up (by the aspect ratio being 2.3 or above),
+ * it approach the rest of the distance with it in front.
  */
 @Deprecated
-public class ApproachCurve extends Command {
+public class ApproachCurveVerbose extends Command {
 
 	/** PID used for approaching the wall. */
 	private SynchronousPIDF PIDapproach;
@@ -26,7 +25,7 @@ public class ApproachCurve extends Command {
 	private double valueapproach;
 	/** Calculated PID output from {@link #PIDcenter} should stored in value. */
 	private double valuecenter;
-
+	
 	/** Height of the left target, overall target, and right target
 	 * (in that order as left as 0, and right as 2)
 	 */
@@ -66,7 +65,7 @@ public class ApproachCurve extends Command {
 	private double heightRatio;
 
 	@Deprecated
-	public ApproachCurve() {
+    public ApproachCurveVerbose() {
 		requires(Robot.drivetrain);
 		requires(Robot.camera);
 		requires(Robot.lineFollowers);
@@ -74,22 +73,29 @@ public class ApproachCurve extends Command {
 		valueapproach = 0;
 		valuecenter = 0;
 
-		PIDapproach = new SynchronousPIDF(QuickAccessVars.KP_APPROACH, QuickAccessVars.KI_APPROACH,
-				QuickAccessVars.KD_APPROACH);
-		PIDcenter = new SynchronousPIDF(QuickAccessVars.KP_CENTER, QuickAccessVars.KI_CENTER,
-				QuickAccessVars.KD_CENTER);
+		PIDapproach = new SynchronousPIDF(
+			QuickAccessVars.KP_APPROACH,
+			QuickAccessVars.KI_APPROACH,
+			QuickAccessVars.KD_APPROACH
+		);
+		PIDcenter = new SynchronousPIDF(
+			QuickAccessVars.KP_CENTER,
+			QuickAccessVars.KI_CENTER,
+			QuickAccessVars.KD_CENTER
+		);
 
 		timer = new Timer();
 	}
 
 	@Override
 	protected void initialize() {
-		PIDapproach.setSetpoint(QuickAccessVars.SETPOINT_APPROACH); // Robot should aim to be be <setpoint> in away from the target
-		try {
-			PIDapproach.setOutputRange(-0.7, 0.7);
-		} catch (Exception e) {
-		}
+		//PID.setIzone(minimumI, maximumI);
+		//PIDapproach.setOutputRange(-1, 1);
+		PIDapproach.setSetpoint(QuickAccessVars.SETPOINT_APPROACH); // Robot should aim to be be 50 in away from the target
+		try {PIDapproach.setOutputRange(-0.7, 0.7);}
+		catch(Exception e) {}
 
+		//PIDcenter.setOutputRange(-1, 1);	
 		PIDcenter.setSetpoint(QuickAccessVars.SETPOINT_CENTER); // Robot should aim to keep the target centered on the crosshair
 
 		timer.start();
@@ -97,25 +103,31 @@ public class ApproachCurve extends Command {
 		intendedPipe = PIPERIGHT;
 		Robot.camera.setPipeline(intendedPipe);
 
-		target_height = new double[3];
+		target_height= new double[3];
 		target_x = new double[3];
 		target_distance = 0;
 
 		driver_control = false;
 		data_missing = false;
-	}
 
+		System.out.println("WWDEBUG: initialize() of ApproachCurve");
+		
+	}
+	
 	@Override
 	protected void execute() {
-
+		
 		updateTargetData();
 
 		if (Math.abs(Robot.input.getRightY()) > QuickAccessVars.CAMERA_DRIVE_THRESHOLD
-				|| Math.abs(Robot.input.getRightX()) > QuickAccessVars.CAMERA_DRIVE_THRESHOLD) {
+			||
+			Math.abs(Robot.input.getRightX()) > QuickAccessVars.CAMERA_DRIVE_THRESHOLD)
+		{
 			driver_control = true;
+			System.out.println("WWDEBUG: driver_control = true");
 		}
 
-		if (!driver_control) { // if the drivers are not controlling it do the auto part
+		if(!driver_control) { // if the drivers are not controlling it do the auto part
 			auto();
 		} else { // if the drivers are controlling it then let them do arcade drive
 			drive();
@@ -124,70 +136,92 @@ public class ApproachCurve extends Command {
 
 	/** Moves the robot autonomously to the target. */
 	private void auto() {
-		if (target_height[LEFT] == 0 || target_height[OVERALL] == 0 || target_height[RIGHT] == 0 || target_x[LEFT] == 0
-				|| target_x[OVERALL] == 0 || target_x[RIGHT] == 0 || target_distance == 0) {
+		System.out.println("WWDEBUG: auto()");
+		if (
+			target_height[LEFT] == 0 || target_height[OVERALL] == 0 || target_height[RIGHT] == 0 ||
+			target_x[LEFT] == 0 || target_x[OVERALL] == 0 || target_x[RIGHT] == 0 ||
+			target_distance == 0
+			)
+		{
 			data_missing = true;
+			System.out.println("WWDEBUG: Skipping moving, lacks data.");
 			return; // if any part of the data is unwritten to, then return out of function to avoid the
 			// the robot driving before it has data
 		} else {
 			data_missing = false;
 		}
 
+		debug_missing();
+
 		if (Robot.camera.canSeeObject()) {
+			System.out.println("WWDEBUG: Can see object");
 
 			shiftCenter();
 
+			System.out.println("WWDEBUG: Calculating PIDs");
 			valueapproach = PIDapproach.calculate(target_distance, timer.get());
 			valuecenter = PIDcenter.calculate(target_x[OVERALL], timer.get());
-
+			
 			// helps to keep the robot to drive in a missile approach curve
 		} else {
+			System.out.println("WWDEBUG: Can not see object, setting motion to 0");
 			// Set value to zero if the target can not be seen so robot does not go crazy
-			valueapproach = 0;
+			valueapproach = 0; 
 			// Don't 0 valuecenter because it should "remember" what direction it's attempting to turn.
 			valuecenter = 0;
 		}
+
+		System.out.println("WWDEBUG: -valueapproach : "+Double.toString(-valueapproach));
+		System.out.println("WWDEBUG: -valuecenter : "+Double.toString(-valuecenter));
 		Robot.drivetrain.arcadeDriveRaw(-valueapproach, -valuecenter);
+		//System.out.println(valueapproach + " " + valuecenter);
 	}
 
 	private void drive() {
+		System.out.println("WWDEBUG: drive()");
+		// Robot.camera.setPipeline(CameraSubsystem.PIPELINE_DRIVER); Conner does not want this
 		Robot.drivetrain.arcadeDriveTeleop(Robot.input.getRightY(QuickAccessVars.ARCADE_FORWARD_MODIFIER),
-				Robot.input.getRightX(QuickAccessVars.ARCADE_TURN_MODIFIER));
+        	Robot.input.getRightX(QuickAccessVars.ARCADE_TURN_MODIFIER));
 	}
 
 	private void updateTargetData() {
+		System.out.println("WWDEBUG: updateTargetData()");
 
 		if (Robot.camera.getPipeline() == intendedPipe) { // wait for pipeline to be the intened one
 			switch (intendedPipe) {
-
-			case PIPELEFT: // Left
-				target_height[LEFT] = Robot.camera.getTargetHeight();
-				target_x[LEFT] = Robot.camera.getObjectX();
-				Robot.camera.setPipeline(PIPEOVERALL);
-				intendedPipe = PIPEOVERALL;
-				break;
-			case PIPEOVERALL: // Overall
-				target_height[OVERALL] = Robot.camera.getTargetHeight();
-				target_x[OVERALL] = Robot.camera.getObjectX();
-				target_distance = Robot.camera.getTargetDistance();
-				// don't move off the center pipeline if the targets become centered
-				if (!(1 / 1.1 < heightRatio && heightRatio < 1.1 / 1) || data_missing) {
-					Robot.camera.setPipeline(PIPERIGHT);
-					intendedPipe = PIPERIGHT;
-				} else {
-					// set the targets to be the same height
-					// so it doesn't try to turn to fix it
-					heightRatio = 1;
-				}
-				break;
-			case PIPERIGHT: // Right
-				target_height[RIGHT] = Robot.camera.getTargetHeight();
-				target_x[RIGHT] = Robot.camera.getObjectX();
-				Robot.camera.setPipeline(PIPELEFT);
-				intendedPipe = PIPELEFT;
-				break;
-			default:
-				break;
+		
+				case PIPELEFT: // Left
+					System.out.println("WWDEBUG: Looking at Left");
+					target_height[LEFT] = Robot.camera.getTargetHeight();
+					target_x[LEFT] = Robot.camera.getObjectX();
+					Robot.camera.setPipeline(PIPEOVERALL);
+					intendedPipe = PIPEOVERALL;
+					break;
+				case PIPEOVERALL: // Overall
+					System.out.println("WWDEBUG: Looking at Both");
+					target_height[OVERALL] = Robot.camera.getTargetHeight();
+					target_x[OVERALL] = Robot.camera.getObjectX();
+					target_distance = Robot.camera.getTargetDistance();
+					// don't move off the center pipeline if the targets become centered
+					if ( !(1/1.1 < heightRatio && heightRatio < 1.1/1) || data_missing) {
+						Robot.camera.setPipeline(PIPERIGHT);
+						intendedPipe = PIPERIGHT;
+					} else {
+						// set the targets to be the same height
+						// so it doesn't try to turn to fix it
+						heightRatio = 1;
+						System.out.println("WWDEBUG: Staying looking for Both (assuming straight on)");
+					}
+					break;
+				case PIPERIGHT: // Right
+					System.out.println("WWDEBUG: Looking at Right");
+					target_height[RIGHT] = Robot.camera.getTargetHeight();
+					target_x[RIGHT] = Robot.camera.getObjectX();
+					Robot.camera.setPipeline(PIPELEFT);
+					intendedPipe = PIPELEFT;
+					break;
+				default:
+					break;
 			}
 		}
 
@@ -195,9 +229,12 @@ public class ApproachCurve extends Command {
 			if (heightRatio != 1) {
 				// if the height ratio doesn't say its straight on...
 				// find how off it is
-				heightRatio = target_height[LEFT] / target_height[RIGHT];
+				heightRatio = target_height[LEFT]/target_height[RIGHT];
+				System.out.println("WWDEBUG: Set heightRatio = "+Double.toString(heightRatio));
 			}
-		} catch (Exception e) {
+		} catch(Exception e) {
+			System.out.println(e);
+			System.out.println("WWDEBUG: Error above forces heightRatio to be 0");
 			heightRatio = 0;
 		}
 
@@ -207,51 +244,66 @@ public class ApproachCurve extends Command {
 	 * Shift the setpoint of the {@link #PIDcenter}
 	 * with the intent of making the robot turn to compensate the angle the robot is facing the target at.
 	 * This should make the turn be an arc to make the robot perpendicular to the target.
-	 */
-	private void shiftCenter() {
+	 */ 
+	private void shiftCenter(){
+		System.out.println("WWDEBUG: shiftCenter()");
+
 		// don't move the setpoint if there is an issue
-		if (heightRatio == 0)
-			return;
+		if (heightRatio==0) return;
 
 		// bound the height ratio within 2 to 0.5
 		// this is so the turn is not to large
-		if (heightRatio > 2)
-			heightRatio = 2;
-		if (heightRatio < 0.5)
-			heightRatio = 0.5;
+		if (heightRatio>2) heightRatio=2;
+		if (heightRatio<0.5) heightRatio=0.5;
 
 		// turn the ratio into a screenspace
 		// by reversing a function that looks similar to the output from the ratio
 		// (the function is the exponential R(x)=2^(-x) )
 		// (so when x is -1 (full left), the ratio is 2)
 		// (so when x is 1 (full right), the ratio is 0.5)
-
+		
 		// solving for x we get -log_2(R(x)) = x
 		// there is no log base in java so we must use the property log_b(a) = log(a)/log(b)
 		double x = -Math.log(heightRatio) / Math.log(2);
 
 		// if the x is within tolerable range, ignore it
-		if (Math.abs(x) < 0.1)
-			x = 0;
+		if (Math.abs(x)<0.1) x=0;
 
 		// now with x being between -1 and 1 we can multiply by the degrees we want the camera to turn
 		// at most (25 degrees)
 		double point = x * 25;
+		System.out.println("WWDEBUG: Set PIDcenter's setpoint to: "+Double.toString(point));
 		PIDcenter.setSetpoint(point);
 	}
 
-	@Override
-	protected boolean isFinished() {
-		return false;
+	private void debug_missing() {
+		if( target_height[LEFT] == 0 ) { System.out.println("WWDEBUG: Missing data: target_height[LEFT]"); }
+		if( target_height[OVERALL] == 0 ) { System.out.println("WWDEBUG: Missing data: target_height[OVERALL]"); }
+		if( target_height[RIGHT] == 0 ) { System.out.println("WWDEBUG: Missing data: target_height[RIGHT]"); }
+		if( target_x[LEFT] == 0 ) { System.out.println("WWDEBUG: Missing data: target_x[LEFT]"); }
+		if( target_x[OVERALL] == 0 ) { System.out.println("WWDEBUG: Missing data: target_x[OVERALL]"); }
+		if( target_x[RIGHT] == 0 ) { System.out.println("WWDEBUG: Missing data: target_x[RIGHT]"); }
+		if( target_distance == 0 ) { System.out.println("WWDEBUG: Missing data: target_distance"); }
 	}
-
+	
+    @Override
+	protected boolean isFinished() {
+		// return Robot.lineFollowers.onLine();
+		return false;
+		/*
+		return (Robot.camera.getTargetDistance() < QuickAccessVars.SETPOINT_APPROACH &&
+			PIDapproach.onTarget(QuickAccessVars.TOLERANCE_APPROACH));
+		*/
+	}
+	
 	@Override
 	protected void end() {
+		System.out.println("WWDEBUG: end() of ApproachCurve");
 		timer.stop();
 		PIDapproach.reset();
 		PIDcenter.reset();
-		valueapproach = 0;
-		valuecenter = 0;
+		valueapproach=0;
+		valuecenter=0;
 		Robot.camera.setPipeline(CameraSubsystem.PIPELINE_CENTER);
 		Robot.drivetrain.stopDrive();
 	}
